refactor(auth): migrate jwt strategy to auth
