+++ src/app/auth/auth.module.ts
@@ -6,7 +6,7 @@ import { PassportModule } from '@nestjs/passport'
 import { AuthService } from './auth.service'
 import { AuthController } from './auth.controller'
 import { UserRepository } from './user.repository'
-import { JwtStrategy } from '../../core/strategies'
+import { JwtStrategy } from './strategies/jwt'
 
 @Module({
   imports: [
