--- src/core/strategies/index.ts
@@ -1,1 +1,0 @@
-export { JwtStrategy } from './jwt'
